Bitpacking validity is checked first when getting a scalar
