+++ encodings/fastlanes/src/bitpacking/compute/mod.rs
@@ -1,12 +1,9 @@
-use vortex::compute::unary::{scalar_at, ScalarAtFn};
+use vortex::compute::unary::ScalarAtFn;
 use vortex::compute::{ArrayCompute, SearchSortedFn, SliceFn, TakeFn};
-use vortex::ArrayDType;
-use vortex_error::VortexResult;
-use vortex_scalar::Scalar;
 
-use crate::bitpacking::compress::unpack_single;
 use crate::BitPackedArray;
 
+mod scalar_at;
 mod search_sorted;
 mod slice;
 mod take;
@@ -26,22 +23,5 @@ impl ArrayCompute for BitPackedArray {
 
     fn take(&self) -> Option<&dyn TakeFn> {
         Some(self)
-    }
-}
-
-impl ScalarAtFn for BitPackedArray {
-    fn scalar_at(&self, index: usize) -> VortexResult<Scalar> {
-        if let Some(patches) = self.patches() {
-            // NB: All non-null values are considered patches
-            if self.bit_width() == 0 || patches.with_dyn(|a| a.is_valid(index)) {
-                return scalar_at(&patches, index)?.cast(self.dtype());
-            }
-        }
-
-        if !self.validity().is_valid(index) {
-            return Ok(Scalar::null(self.dtype().clone()));
-        }
-
-        unpack_single(self, index)?.cast(self.dtype())
     }
 }
